Error on <br> tags in source files
They are almost never semantically correct: https://html.spec.whatwg.org/multipage/text-level-semantics.html#the-br-element
