+++ vale/config/scripts/SmartQuotes_test.go
@@ -13,22 +13,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func mustParseMatches(a []interface{}) []map[string]int {
-	matches := []map[string]int{}
-
-	for _, i := range a {
-		m, _ := i.(map[string]any)
-		match := map[string]int{
-			"begin": int(m["begin"].(int64)),
-			"end":   int(m["end"].(int64)),
-		}
-
-		matches = append(matches, match)
-	}
-
-	return matches
-}
-
 func TestSmartQuotes(t *testing.T) {
 	t.Run("don't use", func(t *testing.T) {
 		src := `Don't use:
